Revert "fixed issues with static site"
