+++ components/Nav.js
@@ -6,10 +6,10 @@ import SocialIcon from './SocialIcon'
 const Nav = () => {
     return (
         <nav className='grid grid-rows-3 p-5 pb-7 bg-custom-background'>
-            <div className='pb-1 row-span-1 flex justify-center text-4xl font-semibold text-custom-text-primary'>
+            <div className='pb-1 row-span-1 flex justify-center text-2xl font-semibold text-custom-text-primary'>
                 D. LOGAN KISER
             </div>
-            <div className='pb-1 row-span-1 flex justify-center text-2xl font-light text-custom-text-secondary'>
+            <div className='pb-1 row-span-1 flex justify-center text-xl font-light text-custom-text-secondary'>
                 <div className='mr-5'>
                     <Link href='/'>
                         <a className="text-xl">Portfolio</a>
@@ -28,7 +28,7 @@ const Nav = () => {
             </div>
             <div className='row-span-1 flex justify-center text-xl font-light'>
                 <div className='mr-1'>
-                        <a href='https://www.linkedin.com/in/logan-kiser/' className="text-xl">
+                        <a href='https://www.linkedin.com/in/logan-kiser-6a103b92/' className="text-xl">
                             <SocialIcon source='/icons/linkedin.png' />
                         </a>
                 </div>
