+++ components/ProjectCard.js
@@ -1,45 +1,24 @@
-// library
-import Image from 'next/image'
-import { useRouter } from 'next/router'
-import { useState, useEffect } from 'react'
-// component
 import TechIcon from "./TechIcon";
 
-const ProjectCard = ({ projectId }) => {
-    const [projectData, setProjectData] = useState({});
-    const [technologies, setTechnologies] = useState([])
-
-    useEffect(() => {
-        fetch(`/api/getProjectById/${projectId}`).then((queryResult) => {
-            queryResult.json().then((projectObject) => {
-                setProjectData(projectObject.project.Item);
-            })
-        })
-    }, []);
-
-    const router = useRouter()
-
+const ProjectCard = ({ project, objectFit, objectScaling='object-none' }) => {
+    let imgStyle = objectScaling + ' ' + objectFit + ' ' + 'rounded-md + max-h-40 max-w-40';
   return (
-      <div className='w-3/4 max-w-xl max-h-fit p-2 mb-4 flex self-center bg-custom-card rounded-md hover:bg-custom-hover' onClick={() => router.push('/projectDetails/' + projectId)}>
+      <div className='w-3/4 max-w-md max-h-sm p-2 mb-4 flex self-center bg-custom-card rounded-md hover:bg-custom-hover'>
             <div className='flex'>
-                <div className='w-32 h-32 shrink-0 relative'>
-                    <Image 
-                    src={(!projectData.snippet) ? '/icons/react.png' : projectData.snippet}
-                    layout='fill'
-                    objectFit='contain'
-                    />
+                <div className='w-22 shrink-0'>
+                    <img className={imgStyle} src={project.snippet} alt=''/>
                 </div>
                 <div className='self-start'>
-                    <div className='projectNameDiv'>
-                        {projectData.name}
+                    <div className='pt-1 pl-6 text-2xl font-bold text-custom-text-primary'>
+                        {project.name}
                     </div>
-                    <div className='pl-6 flex flex-wrap relative'>
-                        {projectData.technologies ? projectData.technologies.map((technology) => {
+                    <div className='pl-6 flex flex-wrap'>
+                        {project.technologies.map((technology) => {
                             return (
-                                <TechIcon key={technology.id} source={technology.icon} />
+                                <TechIcon key={technology.name} source={technology.icon} />
                             )
-                        }) : 'Loading'}
-                        <a href={projectData.repo}>
+                        })}
+                        <a href={project.repo}>
                             <TechIcon source='/icons/github.png' />
                         </a>
                     </div>
@@ -52,8 +31,8 @@ const ProjectCard = ({ projectId }) => {
                     {/* <div className='pl-6 text-md text-left text-custom-text-secondary hover:text-custom-text-primary'>
                         <a href={project.repo}>GitHub Repo</a>
                     </div> */}
-                    <div className='pl-6 pt-2 text-base font-light text-left text-custom-text-secondary'>
-                        {projectData.description}
+                    <div className='pl-6 pt-3 text-xs font-light text-left text-custom-text-secondary'>
+                        {project.description}
                     </div>
                 </div>
             </div>
