+++ components/SocialIcon.js
@@ -1,15 +1,7 @@
-// library
-import Image from 'next/image'
 
 const SocialIcon = ({ source }) => {
   return (
-    <div className='mr-1 mb-1 w-9 h-9 relative'>
-      <Image
-      src={source}
-      layout='fill'
-      objectFit='contain'
-      alt='' />
-    </div>
+    <img className='pr-1 mb-1 max-h-6 max-w-6' src={source} alt='' />
   )
 }
