+++ components/TechIcon.js
@@ -1,17 +1,8 @@
-// library
 import React from 'react';
-import Image from 'next/image'
 
 const TechIcon = ({ source }) => {
   return (
-
-    <div className='mr-1 mb-1 w-10 h-10 relative'>
-      <Image
-      src={source}
-      layout='fill'
-      objectFit='contain'
-      />
-    </div>
+    <img className='pr-1 mb-1 max-h-6 max-w-6' src={source} alt='' />
   )
 }
