+++ next.config.js
@@ -1,9 +1,14 @@
+const withPlugins = require('next-compose-plugins');
+const optimizedImages = require('next-optimized-images');
 
-module.exports = {
-  distDir: '.next',
+module.exports = withPlugins([
+  [optimizedImages, {
+    /* config for next-optimized-images */
+    handleImages: ['jpg', 'jpeg', 'png'],
+  }],
+
+  // your other plugins here
+
+], {
   reactStrictMode: true,
-  // to play nicely with docker
-  experimental: {
-    outputStandalone: true,
-  },
-}
+});
