+++ package.json
@@ -8,12 +8,13 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "aws-sdk": "^2.1175.0",
+    "imagemin-mozjpeg": "^9.0.0",
+    "imagemin-optipng": "^8.0.0",
     "next": "^12.0.8",
+    "next-compose-plugins": "^2.2.1",
+    "next-optimized-images": "^2.6.2",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "serverless-mysql": "^1.5.4",
-    "sharp": "^0.30.7"
+    "react-dom": "^17.0.2"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.2",
