--- pages/api/getProjectIds.js
@@ -1,7 +1,0 @@
-
-// import db from '../../aws/dbcon'
-
-// export default async function handler (req, res) {
-//     const queryResult = await db.query('SELECT id FROM projects');
-//     res.status(200).json({ records : queryResult.map(elt => elt["id"]) });
-// }
