+++ styles/globals.css
@@ -5,7 +5,3 @@
 .imgBorder {
   @apply rounded-md;
 }
-
-.projectNameDiv {
-  @apply pl-6 text-3xl font-bold text-custom-text-primary;
-}
